add support for `require_chef_omnibus` driver option
This option works just like it does in `kitchen-ec2`. It ensures the 
desired version of Chef is installed via the official Opscode Omnibus 
client package. 

The main reason for adding this is to allow projects to use the official 
Canonical published Vagrant base boxes: 

http://uec-images.ubuntu.com/vagrant/

These base boxes do not ship with Chef pre-installed.