[Breaking] Support Vagrant 1.1+ and remove vagrant gem dependency.
A new strategy is used here when dealing with Vagrant: each Kitchen
Instance gets its own dedicated Vagrantfile which is rendered out just
before any Vagrant CLI commands are invoked. In this way we can avoid
writing a vagrant plugin and we return any Vagrantfile in the project
root back to the user to fully control.

An isolated Vagrant sandbox is created for each Instance in a an
internal directory under .kitchen/. For example, given an Instance name
of "default-ubuntu-1204", a Vagrantfile will be created in:

  .kitchen/kitchen-vagrant/default-ubuntu-1204/Vagrantfile