[Bugfix] Calling #destroy should not memoize #create_vagrantfile.
This edgecase happens when a instance already exists and `kitchen test`
is called on that instance. Since the test action always calls destroy,
and destroy cleans up the vagrant root directory, we want to re-create
the directory and Vagrantfile when the create action is invoked.

Super thanks to @sandfish8 for helping to reproduce!!!