Qualify VM names with cookbook name
VirtualBox names VMs after the base name of the current directory, which
results in names like default-centos-65_default_..., which are not very
helpful when looking at a process list or working in the VirtualBox GUI,
especially when working with several cookbooks at once.

Change the working directory name created by kitchen-vagrant to look
like

    kitchen-${cookbook}-${rest as before}

to make things clearer.  (${cookbook} is actually just the name of the
root directory for Test Kitchen, but that's usually a cookbook name.)