Use the vagrant-winrm Vagrant plugin to resolve VM IP address.
It turns out that the IP address of the Vagrant VM returned by the
`vagrant ssh-config` varies with each Vagrant provider and so can't be
relied on to return a private/nat'ed address. In some cases
`"127.0.0.1"` is returned and the locally forward port is returned for
the port number.

To resolve this, the `vagrant-winrm` Vagrant plugin will be used,
exposing a `vagrant winrm-config` subcommand to return WinRM-specific
connection details of the form:

    Host CustomHostname
      HostName Windows2008VM.vagrant.up
      Port 5985
      User vagrant
      Password vagrant

This plugin requires version 1.6 of Vagrant or higher and this Driver
will check to see that the plugin is installed and the Vagrant version
is sufficient--if not a UserError is raised with a message describing
what action(s) to take.

If a WinRM transport is not used, then the minimum version of Vagrant
remains 1.1.0 (as before) and the `vagrant-winrm` plugin will not be
required keeping a slimer set of requirements for users who don't care
to manage Windows VMs.

For more details about the `vagrant-winrm` plugin, please see:

https://github.com/criteo/vagrant-winrm

Closes #153