Allow a fuzzier match for known Bento box names.
This update attempts to improve on
a2ef401fb5a644c96165ce836ff3e6379637a37f which remains a workaround
until the work of continuously publishing Bento boxes to Atlas is
completed.

The rationale and rules in the above mentioned commit remain, but now
only the platform name is checked. In other words, if the platform name
begins with one of the following (case-sensitive):

* `centos`
* `debian`
* `fedora`
* `freebsd`
* `opensuse`
* `ubuntu`

then the platform is assumed to have a Bento box, hosted in the
pre-determined S3 bucket.

References #165
References test-kitchen/test-kitchen#707
References test-kitchen/test-kitchen#663
References chef/chef-dk#382
References chef/chef-dk#378