shorten the directory name for vagrant_root
This omits "kitchen-#{something}" from the vagrant root directories that
will appear under .kitchen/kitchen-vagrant. 

Windows users will run into a path length limit of 260 characters.[1]
We have been repeating in the directory structure "kitchen-" and the
basename of kitchen-root--the directory that contains the .kitchen.yml
which most of the time works out to be the name of the cookbook being
tested. This made for meaningful if verbose names for virtual machines
so as to correlate a machine running in a Vagrant-supported hypervisor
to a test-kitchen config.

The inclusion of this string was _not_ for any unique identifier to
avoid naming collusions within the hypervisor. This change should be
safe for an upgrade for kitchens that have running instances prior to
upgrade. The state file for a machine that was successfully created will
have recorded the vagrant_root path used at creation time. This will
allow kitchen commands to continue to operate on the running instances.

[1] https://github.com/test-kitchen/kitchen-vagrant/issues/210

Signed-off-by: Robb Kidd <robb@thekidds.org>