disable audio by default with VirtualBox guests
Fixes https://github.com/test-kitchen/kitchen-vagrant/issues/388

As mentioned in that issue, modifying the VM to have audio set to "none"
will disable sound for the VM and prevent VirtualBox from connecting to
the host's audio which really cramps a user's style when they're
listening to something over bluetooth headphones.

This change opts to be simple-but-repetitive in the tests to include the
new default in each example VM config.

Signed-off-by: Robb Kidd <robb@thekidds.org>