resolve deprecation notice about Fixnum
Fixnum was deprecated in Ruby 2.4, so this should be checking for
Integer.

Changed this check to `kind_of?` for backwards compatibility. Fixnum is
a subclass of Integer, even in Ruby < 2.4. `kind_of?` will return true
if the class given is the class of the object, or if it is one of the
superclasses of the object or modules included in the object.
`instance_of?` only returns true for an exact match on the class, does
not take subclassing into consideration.

Signed-off-by: Robb Kidd <robb@thekidds.org>