Hyper-v Name Length fix
As part of https://github.com/test-kitchen/kitchen-vagrant/commit/b018a077b93cc343f29c43e9eec2366f38eefe61 to introduce unique names a uuid was added to the generated vagrantfile. This causes problems with hyper-v which relies on names being 100 characters or less otherwise it throws an exception.

This commit takes the current naming system and substrings it to fit
within this character limit. It may be a better idea to use time ticks
as these would end up with us substringing less due to their shorter
nature (the UUID is a third of the allowed length by itself)

1.5.2

2.2.5

2.2.5

Windows 10

Signed-off-by: Xorima <jason@avon-lea.co.uk>