feat: support setting extra data for VirtualBox (#493)
VirtualBox supports supplying VMs with additional data using the
`setextradata` command in `VBoxManage`. Setting data here is sometimes
necessary to get guests to boot on unfamiliar systems (such as by
telling the guest not to try to read SMC data from the host).

As this is VirtualBox-specific, it's implemented using a magic value
to the `customize` block.

Co-authored-by: Corey Hemminger <hemminger@hotmail.com>